fix(#31) updated schema bug
