+++ option_bot/db_tools/schemas.py
@@ -217,8 +217,8 @@ class OptionsQuotesRaw(Base):
     bid_exchange = Column(Integer)
     bid_price = Column(DECIMAL(19, 4), nullable=False)
     bid_size = Column(Integer, nullable=False)
-    sequence_number = Column(Integer)
-    sip_timestamp = Column(Integer, nullable=False)
+    sequence_number = Column(BigInteger)
+    sip_timestamp = Column(BigInteger, nullable=False)
     created_at = Column(DateTime, server_default=func.now())
     updated_at = Column(DateTime, server_default=func.now(), onupdate=datetime.now(dt.UTC))
     is_overwritten = Column(Boolean, server_default=expression.false())
